requirements.txt
Google-Images-Search==1.4.6
python-dotenv==1.0.0
windows-curses==2.3.1